# ARROW-737: [C++] Enable mutable buffer slices, SliceMutableBuffer function