@wesm 2017-04-02
On a faster computer (i7-6700K @ 4.2 GHz) things are less conclusive (maybe ~1% speedup). I think we'd need to plot the distribution to see the median and variance for these microsecond-level measurements

before:

```
Run on (8 X 4200.16 MHz CPU s)
2017-04-02 12:00:39
Benchmark                                                  Time           CPU Iterations
----------------------------------------------------------------------------------------
BM_WriteRecordBatch/1/min_time:1.000/real_time         38996 ns      38998 ns      35949   25.0424GB/s
BM_WriteRecordBatch/4/min_time:1.000/real_time         40939 ns      40940 ns      34171   23.8543GB/s
BM_WriteRecordBatch/16/min_time:1.000/real_time        34525 ns      34526 ns      40565    28.286GB/s
BM_WriteRecordBatch/64/min_time:1.000/real_time        37080 ns      37081 ns      37613   26.3369GB/s
BM_WriteRecordBatch/256/min_time:1.000/real_time       57401 ns      57403 ns      24382    17.013GB/s
BM_WriteRecordBatch/1024/min_time:1.000/real_time     146338 ns     146343 ns       9093   6.67334GB/s
BM_WriteRecordBatch/4k/min_time:1.000/real_time       493599 ns     493615 ns       2837   1.97845GB/s
BM_WriteRecordBatch/8k/min_time:1.000/real_time       919002 ns     919040 ns       1509   1088.14MB/s
BM_ReadRecordBatch/1/min_time:1.000/real_time            517 ns        517 ns    2710222    1.8463TB/s
BM_ReadRecordBatch/4/min_time:1.000/real_time           1154 ns       1154 ns    1215516   846.004GB/s
BM_ReadRecordBatch/16/min_time:1.000/real_time          3707 ns       3707 ns     377487   263.419GB/s
BM_ReadRecordBatch/64/min_time:1.000/real_time         15705 ns      15706 ns      89302   62.1798GB/s
BM_ReadRecordBatch/256/min_time:1.000/real_time        61675 ns      61678 ns      22750   15.8339GB/s
BM_ReadRecordBatch/1024/min_time:1.000/real_time      244899 ns     244908 ns       5718   3.98762GB/s
BM_ReadRecordBatch/4k/min_time:1.000/real_time        985370 ns     985409 ns       1421   1014.85MB/s
BM_ReadRecordBatch/8k/min_time:1.000/real_time       1981668 ns    1981745 ns        707   504.625MB/s
```

after:

```
Run on (8 X 4200 MHz CPU s)
2017-04-02 12:01:36
Benchmark                                                  Time           CPU Iterations
----------------------------------------------------------------------------------------
BM_WriteRecordBatch/1/min_time:1.000/real_time         38927 ns      38929 ns      35898   25.0869GB/s
BM_WriteRecordBatch/4/min_time:1.000/real_time         41042 ns      41044 ns      34152   23.7941GB/s
BM_WriteRecordBatch/16/min_time:1.000/real_time        36929 ns      36931 ns      40226    26.444GB/s
BM_WriteRecordBatch/64/min_time:1.000/real_time        37200 ns      37202 ns      37679   26.2514GB/s
BM_WriteRecordBatch/256/min_time:1.000/real_time       57553 ns      57555 ns      24303   16.9681GB/s
BM_WriteRecordBatch/1024/min_time:1.000/real_time     146361 ns     146367 ns       9538   6.67229GB/s
BM_WriteRecordBatch/4k/min_time:1.000/real_time       490051 ns     490070 ns       2854   1.99278GB/s
BM_WriteRecordBatch/8k/min_time:1.000/real_time       912871 ns     912909 ns       1533   1095.44MB/s
BM_ReadRecordBatch/1/min_time:1.000/real_time            515 ns        515 ns    2716639   1.85105TB/s
BM_ReadRecordBatch/4/min_time:1.000/real_time           1148 ns       1148 ns    1226567   850.878GB/s
BM_ReadRecordBatch/16/min_time:1.000/real_time          3653 ns       3654 ns     383242   267.297GB/s
BM_ReadRecordBatch/64/min_time:1.000/real_time         15582 ns      15582 ns      89795   62.6734GB/s
BM_ReadRecordBatch/256/min_time:1.000/real_time        60900 ns      60902 ns      22838   16.0356GB/s
BM_ReadRecordBatch/1024/min_time:1.000/real_time      245004 ns     245014 ns       5714    3.9859GB/s
BM_ReadRecordBatch/4k/min_time:1.000/real_time        983053 ns     983094 ns       1418   1017.24MB/s
BM_ReadRecordBatch/8k/min_time:1.000/real_time       1970884 ns    1970960 ns        711   507.387MB/s
```