@wesm 2017-04-02
here's the appveyor build: https://ci.appveyor.com/project/wesm/arrow/build/1.0.62. merging